+++ src/AvroConvert/DeserializeByLine/Decoder.cs
@@ -1,9 +1,8 @@
 ﻿using System.IO;
 using System.Linq;
-using SolTechnology.Avro.Codec;
-using SolTechnology.Avro.Constants;
 using SolTechnology.Avro.Exceptions;
-using SolTechnology.Avro.Helpers;
+using SolTechnology.Avro.FileHeader;
+using SolTechnology.Avro.FileHeader.Codec;
 using SolTechnology.Avro.Read;
 using SolTechnology.Avro.Schema;