+++ src/AvroConvert/GetSchema/HeaderDecoder.cs
@@ -1,9 +1,8 @@
 ﻿using System.IO;
 using System.Linq;
 using System.Text;
-using SolTechnology.Avro.Constants;
 using SolTechnology.Avro.Exceptions;
-using SolTechnology.Avro.Helpers;
+using SolTechnology.Avro.FileHeader;
 using SolTechnology.Avro.Read;
 
 namespace SolTechnology.Avro.GetSchema