+++ tests/AvroConvertTests/Unit/FileHeader/CodecTests.cs
@@ -1,0 +1,22 @@
+﻿using SolTechnology.Avro.FileHeader.Codec;
+using Xunit;
+
+namespace AvroConvertTests.Unit.FileHeader
+{
+    public class CodecTests
+    {
+        [Fact]
+        public void CreateCodecFromString_NonExistingString_DefaultCodecIsReturned()
+        {
+            //Arrange
+
+
+            //Act
+            var result = AbstractCodec.CreateCodecFromString("NonExistingCodec");
+
+
+            //Assert
+            Assert.IsType<NullCodec>(result);
+        }
+    }
+}